Move fastapi routes into routes folder
